(#97) Move universal serialization functions to separate header file
